Fix exception logic to preserve partition/mount vars
For some reason, `Mount-DiskImage` passthru will sometimes have a null value for the `DiskNumber` property, so this is piped into `Get-DiskImage`  to preserve all properties. 

Partition information (`$partinfo`) would fail silently, because when `Get-Partition -DiskNumber $mountedDisk.Number -ErrorAction Stop` only returns `$null`, but does not register as an exception, so the `catch {}` blocks would never fire. Instead, this is now partially handled through if statements.

Without these changes, in certain cases the script would execute, but not actually shrink any disks, and it would output that it could not get partition information in the logfile.